+++ Sources/ApiKit/ApiResult.swift
@@ -15,14 +15,13 @@ import Foundation
 public struct ApiResult {
 
     public init(
-        data: Data?,
-        response: URLResponse?
+        data: Data,
+        response: URLResponse
     ) {
         self.data = data
         self.response = response
     }
 
-    public var data: Data?
-    
-    public var response: URLResponse?
+    public var data: Data
+    public var response: URLResponse
 }